returned old dictionary
