+++ .github/workflows/deploy.yml
@@ -53,6 +53,6 @@ jobs:
               OPENAI_API_KEY=${{ secrets.OPENAI_API_KEY }} LANGCHAIN_API_KEY=${{ secrets.LANGCHAIN_API_KEY }} LANGCHAIN_ENDPOINT=${{ secrets.LANGCHAIN_ENDPOINT }} \
               LANGCHAIN_TRACING_V2=${{ secrets.LANGCHAIN_TRACING_V2 }} LANGCHAIN_PROJECT=${{ secrets.LANGCHAIN_PROJECT }} TAVILY_API_KEY=${{ secrets.TAVILY_API_KEY }} \
               STORAGE_ACCESS_KEY=${{ secrets.STORAGE_ACCESS_KEY }} STORAGE_ACCESS_KEY_ID=${{ secrets.STORAGE_ACCESS_KEY_ID }} \
-              STORAGE_ID=${{ secrets.STORAGE_ID }} STORAGE_NAME=${{ secrets.STORAGE_NAME }} \
+              STORAGE_ID=${{ secrets.STORAGE_ID }} STORAGE_NAME=${{ secrets.STORAGE_NAME }} DICTIONARY_API_URL=${{ secrets.DICTIONARY_API_URL }} \
               docker-compose up -d
             sudo docker image prune -f
