+++ src/app/api/translate/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 
 import apiErrorService from '@/services/apiErrorService'
-import translateService from '@/services/translateService'
+import translateAIService from '@/services/translateAIService'
 import { ApiErrType } from '@/types/ErrorTypes'
 
 export async function POST(req: NextRequest):
@@ -9,7 +9,7 @@ export async function POST(req: NextRequest):
   try {
     const { word, inputLanguage, outputLanguage } = await req.json()
 
-    return NextResponse.json(word ? await translateService(word, inputLanguage!, outputLanguage!) : [])
+    return NextResponse.json(word ? await translateAIService(word, inputLanguage!, outputLanguage!) : [])
   } catch (error) {
     const err = error as Error & ApiErrType
